requirements.txt
djangorestframework==3.11.0
django-filter==2.3.0
django==3.0.7
djongo==1.3.3
pymongo==3.10.0
sqlparse==0.2.4
dnspython=1.16.0
